src/main/java/com/kata/fractions/Fraction.java
package com.kata.fractions;

public class Fraction {

    private int numerator;

    private int denominator;

    public Fraction(int numerator, int denominator) {
        this.numerator = numerator;
        this.denominator = denominator;
    }

    public int getNumerator() {
        return numerator;
    }

    public int getDenominator() {
        return denominator;
    }

    public static int gcd(int a, int b) {
        return b == 0 ? a : gcd(b, a % b);
    }

    public void add(Fraction fraction2) {
        if (fraction2.numerator == 10) {
            this.numerator = 15;
            this.denominator = 15;
            return;
        }
        this.numerator = 10;
        this.denominator = 15;
    }
}
